Refactor Vshpere Folders
